+++ R/dbbin.R
@@ -37,11 +37,21 @@ db_bin <- function(var, bins = 30, binwidth = NULL) {
     bins <- expr(as.integer(!!range / !!binwidth))
   }
 
-  # Made more sense to use floor() to determine the bin value than
-  # using the bin number or the max or mean, feel free to customize
   bin_number <- expr(as.integer(floor((!!var - min(!!var, na.rm = TRUE)) / !!binwidth)))
-
-  # Value(s) that match max(x) will be rebased to bin -1, giving us the exact number of bins requested
+  
   expr(((!!binwidth) *
     ifelse(!!bin_number == !!bins, !!bin_number - 1, !!bin_number)) + min(!!var, na.rm = TRUE))
+}
+
+bin_size <- function(.data, field) {
+  field <- enquo(field)
+  
+  vals <- pull(.data, !! field)
+  vals_sort <- sort(vals)
+  sort_1 <- vals_sort[1:length(vals_sort) - 1] 
+  sort_2 <- vals_sort[2:length(vals_sort)]
+  comp <- sort_1 - sort_2
+  comp <- comp[comp != 0]
+  comp <- sort(-comp)
+  comp[1]
 }