+++ tests/testthat/test-raster.R
@@ -19,3 +19,18 @@ test_that("The correct number of rows are returned when using complte", {
     80
   )
 })
+
+test_that("Compute raster 2 returns the rignt number of rows", {
+  expect_equal(
+    nrow(db_compute_raster2(mtcars, wt, mpg, complete = TRUE)),
+    600
+  )
+  expect_equal(
+    nrow(db_compute_raster2(mtcars,
+                           wt, mpg,
+                           complete = TRUE,
+                           resolution = 10
+    )),
+    80
+  )
+})